Add gamefiles and arrange a directory structure
I have first created a GameContent folder and segregated content into 3 parts.
The 3 parts are maps , blueprints and assets .I have added the respective
content to the respective folder .the second file that is present in this
commit is the mannequin file containing some skeletal and animation for it.
The folders that i have not included but should be present in your directory
are the startercontent directory and the AnimStarterPack .
This commit also adds the static camera overlooking the level .